Added a more complex function to compare current object against original object.
This is not very reliable either because certain times we save objects inside of other objects, which makes them compare false but they have not changed.

For example, when we get the vendors rate sheets, we assign them to vendor.rate_sheets… this makes for streamlined data but complicates things because now angular things the object has changed, even though we make separate API calls to save this action and reference - i.e.: its not actually something we need to save on the vendor.