Fix issue with transclusion and parent / child scope
This fixes an issue where the `quoteCost` variable became a property of the child scope. This happens when a directive uses transclude or creates its own isolate scope, as with `scope: {}`.

Variables that are functions or objects remain in the proper scope, so `quote. quoteCost` would work fine. While quoteCost does not. Specifying `$parent. quoteCost` tells the transcluded directive to reference the $parent scope.

@see http://stackoverflow.com/a/14049482/1738217
and @see http://stackoverflow.com/questions/16653004/confused-about-angularjs-transcluded-and-isolate-scopes-bindings